// tmp_app/Api.ts
export interface input$api$auth$update_info {
    f_name: string, l_name: string, username: string, email: string, channel: string
}
// ...
export interface output$api$auth$update_info {
    f_name: string, l_name: string, username: string, email: string, channel: string
}